# lib/api-zod/src/generated/types/taskInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TaskFlow API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TaskInputStatus } from "./taskInputStatus";

export interface TaskInput {
  title: string;
  description?: string;
  /** @nullable */
  assigned_to?: number | null;
  /** @nullable */
  due_date?: string | null;
  status?: TaskInputStatus;
}